+++ CHANGELOG.md
@@ -11,6 +11,7 @@ nav_order: 1
 
 ## [MAJOR.MINOR.PATCH] - YYYY-MM-DD
 
+- Add `is_super_admin` flag to `aiven_organization_application_user` resource
 - Fix `ip_filter` conversion issue
 - Fix `aiven_organization_application_user_token` crashes with empty optional fields
 - Add `aiven_mirrormaker_replication_flow` replication factor